+++ app/models/pay/charge.rb
@@ -15,6 +15,8 @@ module Pay
     validates :amount, presence: true
     validates :processor_id, presence: true, uniqueness: {scope: :customer_id, case_sensitive: true}
 
+    ensure_store :data
+
     # Store the payment method kind (card, paypal, etc)
     store_accessor :data, :paddle_receipt_url
     store_accessor :data, :stripe_account