+++ app/models/pay/subscription.rb
@@ -20,6 +20,8 @@ module Pay
     # Callbacks
     before_destroy :cancel_if_active
 
+    ensure_store :data
+
     store_accessor :data, :paddle_update_url
     store_accessor :data, :paddle_cancel_url
     store_accessor :data, :paddle_paused_from