+++ test/models/pay/merchant_test.rb
@@ -1,7 +1,8 @@
 require "test_helper"
 
 class Pay::MerchantTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
+  test "text data column" do
+    pay_merchants(:one).update!(onboarding_complete: true)
+    assert_equal true, pay_merchants(:one).onboarding_complete
+  end
 end