+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Ensure that `data` columns are `ActiveRecord::Store` with `text` columns. - @excid3
+
 ### 3.0.12
 
 * Add `CHECKOUT_SESSION_ID` to checkout URLs by default and document how to add them - @excid3